Author a path into a running world (#142)
Issue #142 §13 step 6: the network becomes rebuildable, driven by a hook
and a console command and no UI at all. The issue calls this the step
that carries the risk, and it is right — everything about this world is
planned before the first chunk exists on purpose, so that no chunk
carrying a corridor is ever built while roads are being routed.

`World.addPath` is the one hole in that. It routes and profiles exactly
as the planner does — a few hundred height queries, about one chunk of
terrain — adds the edge, rebuilds the spatial index, draws a ribbon for
it, and drops every chunk so the streamer bakes them against a height
field that now has the corridor in it. That last part is `rebuildProps`,
already documented as a TUNING path costing about what walking into fresh
ground costs.

AND THE GROUND UNDER THE HERO MOVES. `world/index.ts` says of
`rebuildProps` that "the hero cannot fall through the result — getHeight
is a pure function of the seed". That sentence is now false for this one
caller: `getHeight` consults `terrain.roads`, which this mutates, and
`carveAt` shifts a column by up to 1.62. `refitHero` in main.ts puts him
back, upward only — a hero in the air when a path is carved under him
should fall, which is what the sim is for. The caller does it because
this module does not own the player, and while mounted his position is
written from the saddle every slice.

THREE REFUSALS, REPORTED (§12f). An unknown profile name lists the ones
that exist; two ends under 12 units apart is not a path; and a profile
that cannot bridge may not be routed into water. That third one is the
first thing the feature actually needed. `PathProfile.bridges` tells the
router to pay a lake's price per wet step, which keeps a footpath out of
water it can go round — but it cannot help when an END is wet, because
the walk has to arrive. `profileRoad` then floors every wet sample to
WATER_LEVEL + 1.9 and flags it a span, and what came back was a footpath
standing 1.9 units over a lake with the carve switched off under it and
no piers to put there. A ford is the third answer and it is not built;
until it is, say so.

Two smaller things the first authored path found:

  - A FREE END HAS TO BE ANCHORED. `profileRoad` takes NaN for "whatever
    the profile says", which is right for the first road out of a town —
    it is what lets the road decide the town's height — and wrong for a
    path that simply stops in open country: the limiter fills dips by
    raising, so the deck arrives wherever the smoothing left it. Anchored
    to `getHeight` at the ROUTE'S OWN ends, not at the ones that were
    asked for: `routeRoad` resamples at SEG_LEN and lets the tail land up
    to a unit away, and on ground that steps in whole units a unit of
    horizontal slop is two units of height.
  - THE REGISTRY WAS A SNAPSHOT. `TownRegistry.roads` is what a compass,
    a signpost and every probe mean by "the roads", taken when the world
    was planned — correct for exactly as long as the network could not
    change. It grows now (`MutableTownRegistry.addRoad`), or
    `__dbgTowns().roads` describes a world with one fewer path than the
    one on the ground.

`segRoad` is Uint16 rather than Uint8. Three roads became twenty-six when
the beaten tracks joined the network and this can add more; 255 is not a
bound anyone would notice being crossed, because the index would wrap and
hand a segment to the wrong road's trim planes.

`tools/test-path-edit.mjs` is the guard, and the assertion that matters
is the third: the cross-section over the AUTHORED corridor, rim to rim,
is issue #15's own measurement pointed at ground that was already meshed
when the path arrived. 0 of 492 samples have terrain over the ribbon, the
deck steps 0.027 against MAX_STEP_UP 0.5, and the hero is not inside the
ground. An editor that can author a path can author one that reopens #15.

No collider changed.

+++ src/world/roads.ts
@@ -439,6 +439,12 @@ export class RoadNetwork implements RoadField, RoadClearance {
    * Which road owns each segment, and a one-entry-per-road cache of "is this
    * query point past that road's trim planes".
    *
+   * Uint16 and not Uint8: three roads became twenty-six when a settlement's
+   * beaten tracks joined the network, and `World.addPath` can add more at
+   * runtime. 255 is not a bound anybody would notice being crossed — the index
+   * would just wrap and hand a segment to the wrong road's trim planes, which
+   * fails silently in the way this comment already warns about below.
+   *
    * PER ROAD, NOT PER SEGMENT, and that is the whole subtlety. Skipping only
    * the terminal segment does not work and fails SILENTLY: a point a tenth of a
    * unit past the end node is still ~3 units from the segment before it, well
@@ -457,7 +463,7 @@ export class RoadNetwork implements RoadField, RoadClearance {
    * interleaved runs each way — before 11.57 / 12.04 / 10.02 ms, after 10.89 /
    * 10.72 / 10.59. The run-to-run spread is larger than the difference.
    */
-  private segRoad = new Uint8Array(0);
+  private segRoad = new Uint16Array(0);
   private clipStamp = new Float64Array(0);
   private clipOut = new Uint8Array(0);
   private queryId = 0;
@@ -508,7 +514,7 @@ export class RoadNetwork implements RoadField, RoadClearance {
     );
     this.seg = new Float32Array(n * 6);
     this.segBridge = new Uint8Array(n);
-    this.segRoad = new Uint8Array(n);
+    this.segRoad = new Uint16Array(n);
     this.roadRole = new Uint8Array(this.roads.length);
     for (let i = 0; i < this.roads.length; i++) {
       const q = this.roads[i].profile.roles;
